requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
psycopg[binary]
pydantic[email]==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9